requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

runpod==1.7.12
torch==2.4.0
torchaudio==2.4.0
ffmpeg-python==0.2.0
requests==2.*
pydantic==2.*
pyannote.audio==3.1.1
numpy>=1.24
scipy>=1.10
